requirements.txt
python-jose[cryptography]
passlib[bcrypt]
fastapi[all]
uvicorn[standard]
sqlalchemy
pymysql
python-multipart
pydantic